+++ oap-server/server-alarm-plugin/src/main/java/org/apache/skywalking/oap/server/core/alarm/provider/discord/DiscordHookCallback.java
@@ -19,103 +19,54 @@
 package org.apache.skywalking.oap.server.core.alarm.provider.discord;
 
 import com.google.gson.JsonObject;
-import io.netty.handler.codec.http.HttpHeaderValues;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.http.HttpHeaders;
-import org.apache.http.HttpStatus;
-import org.apache.http.StatusLine;
-import org.apache.http.client.config.RequestConfig;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.entity.ContentType;
-import org.apache.http.entity.StringEntity;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.util.EntityUtils;
-import org.apache.skywalking.oap.server.core.alarm.AlarmCallback;
 import org.apache.skywalking.oap.server.core.alarm.AlarmMessage;
+import org.apache.skywalking.oap.server.core.alarm.HttpAlarmCallback;
 import org.apache.skywalking.oap.server.core.alarm.provider.AlarmRulesWatcher;
 
 import java.io.IOException;
+import java.net.URI;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Use SkyWalking alarm Discord webhook API.
  */
 @Slf4j
-public class DiscordHookCallback implements AlarmCallback {
-
-    private static final int HTTP_CONNECT_TIMEOUT = 1000;
-    private static final int HTTP_CONNECTION_REQUEST_TIMEOUT = 1000;
-    private static final int HTTP_SOCKET_TIMEOUT = 10000;
+@RequiredArgsConstructor
+public class DiscordHookCallback extends HttpAlarmCallback {
     private final AlarmRulesWatcher alarmRulesWatcher;
-    private final RequestConfig requestConfig;
-
-    public DiscordHookCallback(final AlarmRulesWatcher alarmRulesWatcher) {
-        this.alarmRulesWatcher = alarmRulesWatcher;
-        this.requestConfig = RequestConfig.custom()
-                .setConnectTimeout(HTTP_CONNECT_TIMEOUT)
-                .setConnectionRequestTimeout(HTTP_CONNECTION_REQUEST_TIMEOUT)
-                .setSocketTimeout(HTTP_SOCKET_TIMEOUT)
-                .build();
-    }
 
     /**
      * Send alarm message if the settings not empty
      */
     @Override
-    public void doAlarm(List<AlarmMessage> alarmMessages) {
-        DiscordSettings discordSettings = alarmRulesWatcher.getDiscordSettings();
+    public void doAlarm(List<AlarmMessage> alarmMessages) throws Exception {
+        final var discordSettings = alarmRulesWatcher.getDiscordSettings();
         if (discordSettings == null || discordSettings.getWebhooks().isEmpty()) {
             return;
         }
-        discordSettings.getWebhooks().forEach(webHookUrl -> {
-            alarmMessages.forEach(alarmMessage -> {
-                String content = String.format(
+        for (final var webHookUrl : discordSettings.getWebhooks()) {
+            for (final var alarmMessage : alarmMessages) {
+                final var content = String.format(
                         discordSettings.getTextTemplate(),
                         alarmMessage.getAlarmMessage()
                 );
                 sendAlarmMessage(webHookUrl, content);
-            });
-        });
+            }
+        }
     }
 
     /**
      * Send alarm message to remote endpoint
      */
-    private void sendAlarmMessage(DiscordSettings.WebHookUrl webHookUrl, String content) {
-        JsonObject body = new JsonObject();
+    private void sendAlarmMessage(DiscordSettings.WebHookUrl webHookUrl, String content) throws IOException, InterruptedException {
+        final var body = new JsonObject();
         body.addProperty("username", webHookUrl.getUsername());
         body.addProperty("content", content);
-        sendPostRequest(webHookUrl.getUrl(), body.toString());
+        final var requestBody = body.toString();
+        post(URI.create(webHookUrl.getUrl()), requestBody, Map.of());
     }
 
-    /**
-     * Post rest invoke
-     */
-    private void sendPostRequest(String url, String requestBody) {
-        try (CloseableHttpClient httpClient = HttpClients.createDefault()) {
-            HttpPost post = new HttpPost(url);
-            post.setConfig(requestConfig);
-            post.setHeader(HttpHeaders.ACCEPT, HttpHeaderValues.APPLICATION_JSON.toString());
-            post.setHeader(HttpHeaders.CONTENT_TYPE, HttpHeaderValues.APPLICATION_JSON.toString());
-            StringEntity entity = new StringEntity(requestBody, ContentType.APPLICATION_JSON);
-            post.setEntity(entity);
-            try (CloseableHttpResponse httpResponse = httpClient.execute(post)) {
-                StatusLine statusLine = httpResponse.getStatusLine();
-                if (statusLine != null) {
-                    if (statusLine.getStatusCode() != HttpStatus.SC_OK && statusLine.getStatusCode() != HttpStatus.SC_NO_CONTENT) {
-                        log.error("send to {} failure. Response code: {}, Response content: {}", url,
-                                statusLine.getStatusCode(),
-                                EntityUtils.toString(httpResponse.getEntity())
-                        );
-                    }
-                }
-            } catch (IOException e) {
-                log.error(e.getMessage(), e);
-            }
-        } catch (IOException e) {
-            log.error(e.getMessage(), e);
-        }
-    }
 }