+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/management/ui/template/UITemplateInitializer.java
@@ -62,6 +62,8 @@ public class UITemplateInitializer {
         Layer.VIRTUAL_MQ.name(),
         Layer.AWS_EKS.name(),
         Layer.OS_WINDOWS.name(),
+        Layer.AWS_S3.name(),
+        Layer.AWS_DYNAMODB.name(),
         "custom"
     };
     private final UITemplateManagementService uiTemplateManagementService;